+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 Changelog
 =========
+
 ## Unreleased
 ### Fixed
 - Bug when attempting to resolve a transformer from a Yii config array
+
+## 1.0.0-rc.5 - 2018-07-13
+### Changed
+- Updated Transform library to v3.0
+
+### Removed
+- `Flux::item` and `Flux::collection` are not longer accessible.
 
 ## 1.0.0-rc.4 - 2018-05-11
 ### Changed